Fix "Press return..." message appearing on OS X
Yes, again.

This time fix blocks SIGCHLD signal in my_system() function to ensure
that it won't be handled by appropriate handler in signals.c.  The
handler can not be defaulted, since it will cause loose of SIGCHLD
signals, hence we block it and process afterwards.

Thanks to Daniel R. (aka r1chelt).