Check if the configuration directory exists
The previous file have a bug. `isdirectory('$APPDATA/Vifm')` always
return the same things because the variable is inside the quote.
In Windows I have found that Vifm, save the vimfiles in is directory
if a vifmrc is create.
Otherwhise It create this file into ~/.vifm/ if it exist.
Otherwhise It create inside the $APPDATA/Vifm/ (the directory create by
default).