Add hack to prevent flickering of title line
It's in an ugly hack to avoid flickering of top line of the current
view.  We need attributes colors recalculated correctly before applying
them, otherwise color changes twice on the screen.  Need to generalize this
by updating all color pairs that we can, or that depend on group, whose
properties are changed. */