Provide pair space compression on lack of pairs
TODO: in case of performance issues cache pair_in_use() in the
first loop or change the function to fill in bit field of pairs.

Compression shouldn't be triggered, but is there just in case.

Might also need additional function to redraw interface before
reassigning moved color pairs, but adding one probably not worth
it.