Pull in glibc's getopt_long() implementation
getopt_long() begin GNU extension (I with it to be standardized one day)
is the reason why options were parsed manually till this day.  Turns out
the correct way is to just pull it in and it will compile to nothing if
system uses glibc or will provide proper implementation if it doesn't.