Fix pacman-mode activated via startup commands
vifm --no-configs -c 'set lsview' -c '!find %u'

    . . . . . .
    . . . . . .
    . . . . . .
    . . .

Thanks to filterfalse.

Note that yesterdays fix for the same issue on :restart can be reverted
now, but it seems to be better to keep it to omit max file width change
like N1 -> 0 -> N2 on reload, should be N1 -> N2 at the moment.