Manage cancellation outside of ops_enqueue()
It makes more sense this way and otherwise causes issues in background
operations, raw()/noraw() calls interact with terminal and too many
calls in short period of time seem to leave terminal or curses state
in a wrong state.

This was observed on background operations on directories.

Thanks to Svyatoslav Mishyn (a.k.a. juef).