Fix possible vifminfo shrinking on init error
As if some data were removed from vifminfo.

As an example of wrong behaviour this command due to redirection caused
vifm to abort during initialization and lost most of vifminfo because
'vifminfo' option had default value during merging of vifminfo file:

    ls | vifm --choose-files - > ~/list

I did lost state several times, but was unable to figure out the reason.
This might be it.