Add check for source file name dups in cv
On file operations like copy or move, this could cause undesirable
file overwrite.

Other options include: add suffixes or ask for conflict resolution.
The second one is not deterministic and depend on order of files in
the list, which is confusing especially on move (consider original
file being overwritten by one of two files and now you get confirmation
request for the second file overwriting the first one).  Both options
are not consistent with what's going in already handled similar cases.

Thanks to filterfalse.