Generate BUGS and TODO
From now on these files are snapshots of contents of something akin to
todo-management tool.

New files are a mix of old versions of these files and several other
similar sources.  Eventually, everything should be imported into the
new system, which has proven to be effective, but it's too boring and
time consuming thing to do... hence this incomplete drop, which
nevertheless includes most of tracked TODO records.