Introduce concept of fake and valid entries
A fake entry is one that does not correspond to any file at all
(these are not yet created by current code).

An invalid entry is either ".." or a fake entry.  The common thing
about them is that they do not take part in any file operations nor
can they be selected.

There are quite a few changes, but they are simple.  Mostly fake
entries either abort an operation (when only one element is
processed) or are ignored.  In some cases they are irrelevant and
thus not treated in any special way.  Their properties:
 * can't be selected;
 * aren't updated from file system (and have no size);
 * aren't traversed by iter_* methods;
 * can't be renamed;
 * can't be previewed;
 * their meta-data shouldn't appear anywhere (statusline, file
   info);
 * are of unknown type;
 * can't be opened nor followed;
 * can't be a target of a mark;
 * are displayed as a dotted line in a file view;
 * aren't completed;
 * aren't expanded in macros;
 * aren't inserted on command-line (as if they have no path).

Also disallow selecting ".." entry with :select command (the check
was missing here).

Visual mode contained a mess with regard to processing of ".."
entries, cleaned that up.