Don't do dangerous removal without confirmation
This confirmation can not be skipped.  In some cases instead of
confirmation there will be a warning about excluded files, this
means that they changed their location (were under just moved
directory) or would overwrite just moved directory.  We could
omit the exclusion in second case, but it seems to be helpful to
have it.