Improve comparison by contents algorithm
Now file contents is actually compared when hash of file prefix
matches (256 KiB).  Not hashing whole files helps a lot to speedup
performance of comparison when there are large files.  There is
still a bad case on which performance will degrade, but it requires
having big files of the same size which differ at some positions
after prefix, which isn't that common (but still possible).

Thanks to filterfalse.