Change the way global-local options are handled
Instead of handling it on per-option basis, use newly introduced
universal handler to set values in the other view.

The issue with the old approach (apart from messy code in handlers)
is that if current pane already has value set to the option, handler
doesn't get invoked and other view isn't updated.

Could be seen with 'dotfiles' option stored with different values
in vifminfo and `set dotfiles` in vifmrc: inactive view could be
left unchanged.

Both implementations have the same behaviour for things like += and
similar operators.  In this case both panes get the same value
instead of each pane being processed independently.  Shouldn't be an
issue, but nice to note it.

Thanks to filterfalse.