Escape first tilde in paths that go to the shell
At least sometimes zsh behaves in a strange way and aborts if ~something
doesn't expand (bash just skips expansion in this case).  Also we escape
all such things, so let's be consistent.  Moreover we the motivation of
not escaping it is unclear (not documented).

Do the expansion explicitly for :edit and alike cases, more places might
need to be updated.

The change was introduced with filename modifiers in
e337d814c470f510658059d0ee362322e6e3e8e4.

Thanks to Marius Schmidl.