Fix highlighting of matching in right aligned tree
Not that it's useful, but the correct part of the string should be
highlighted anyway.

The reason this didn't work is that for right aligned columns
invocation of gap filling reset prefix_len making it unavailable
for actual column printing.

The fix is to consider the prefix only for primary column.