Fix backslash in arguments for :*[un]map commands
Required \ to be written as \\.

Started to happen after baeba002dc56591cdcc9664ea5227f5aa8a9bbf4.

Use "raw" arguments of engine/cmds unit for that.  Could be implemented
entirely in cmd_handlers, but that would made as to also validate
number of arguments ourselves, which is responsibility of engine/cmds.

Thanks to filterfalse.