Fix incorrect 'sort' state after :restart
ui_view_sort_list_get() looked suspicious for a while, but it was
decided to do not change it while it works.  Looks like it didn't
really work well though.  Changed it to return local sorting array
in most cases.  All tests still pass and manual tests pass as well,
so assuming it's a real fix rather than a workaround.

Thanks to sudo-nice.

Closes #235 on GitHub.