More precise and correct size formatting
It was possible to get ugly and incorrect formatting by setting
precision to a big number in 'sizefmt'.  For example:
 * 225.-2147483648 K
 * 225.-000000000000002147483648 K

Current results are not perfect (last digits might be non-zero even
when they should or they could be rounded somewhat incorrectly) due
to floating-point arithmetic involved.  `numfmt` doesn't seem to
care about it much too and these are extreme cases anyway.