Make tests added since v0.8.2 "portable"
Where "portable" means that they do not depend on test data and
sandbox to be located at fixed place at absolute path known at
compile-time.  Incorrectly assumed that we don't need this anymore
before, so doing it now...

Also {x} was changed to { x } and PATH_MAX to (PATH_MAX + 1) along
the way.