Revert license of stic to MIT
As was noted in the README of stic before, GPL shouldn't be a problem
for tests, but as it turns out it is if you want to include stic into
your non-GPL project.  This can probably be worked around by extending
GPL with an exception to allow such use case, but just reverting to
MIT will do for now.

Also fix case of the extension.