Extract ui_shutdown() function
Shuts down UI making it possible to use terminal (either after
vifm is closed or when terminal might be used by another
application that vifm runs).

def_prog_mode() doesn't seem to be strictly necessary in all cases,
but in general it looks like it should be called before endwin() to
save current state of the terminal rather than one that was on
initscr().