Fix consuming 100% of CPU by error reading thread
This could happen on certain pattern of running background processes.

Steps to reproduce:
 1. :!sleep 10m &
 2. :!sleep 1 && echo this is a fake error message 1>&2 &
 3. Don't press Enter and CPU load will increase.

The issue was that select() constantly reported file descriptor as
ready after it reached EOF.  And we didn't update set of active
descriptors.

The dialog is tied in by the fact that it prevents job from being
removed, in which case descriptor is closed and read() returns error,
which causes update of list of jobs.

Thanks to petRUShka.

Fixes #296 on GitHub.