Correctly change back current directory after view update
The file browser has the nice ability of changing the current working
directory to the currently displayed one. That changing happens
depending on which view is active, i.e., the left or the right side.
Unfortunately, if there are updates happening on the other side, for
example because a file increased its size, the current working directory
will be changed and not reverted as part of the view update process.
That is a problem when spawning off a new shell and expecting it to have
its working directory set to the currently active view's one.
This change fixes the problem by correctly restoring the current working
directory after each update operation.