Always place cursor on top for newly opened dirs
Otherwise, consider this:

- Open directory `/foo/bar`
- Move cursor to some non-first item
- Type `:cd ..` (so that we'll end up in `/foo`)
- Hit Enter (so that we'll end up in `/foo/bar` again)

At this point, cursor should be on the top. Without the proposed change,
it isn't.

I guess it's a very minor thing for the majority, but I constantly
change directories via external means, so my directory history is
"polluted" by random cursor positions, and it's really annoying when I
open a dir and expect the cursor to be on top, while it isn't.