Report missing quote for "" expression
The idea probably was that comment-only expression is invalid, it's not
a string literal, but this actually only confuses.  Do say that quote is
missing, this also matches Vim's behaviour.  One way to think about it:
comments at the beginning of expressions aren't allowed, so it is in
fact a string literal.