Fix visual mode message not being updated on av
From "-- VISUAL --" to "-- VISUAL (append) --" while in visual mode.

Was appearing only with `set suggestoptions+=visual`.  The reason was
that:
1. It caused an extra update_screen() (in hide_suggestion_box() of
   event_loop.c).
2. Which called print_selected_msg().
3. Which changed curr_stats.save_msg to 2.
4. modes_statusbar_update() was calling only update_vmode_input() and
   not print_selected_msg().

Not updating status bar in visual mode always is probably a mistake, so
change that.  Can't think of a situation when would be in visual mode
and wouldn't need to update status bar to the mode message.

Thanks to Paweł Smolak (a.k.a. psmolak).

Fixes #340 on GitHub.