Try to preserve ownership with 'nosyscalls'
If we aren't root, the error is just ignored.

The original reason why it was --preserve=mode,timestamps without
"owner" is that `cp` would generate "can't preserve owner" on copying
from sshfs as was reported by ranousse in 2012.  This doesn't seem to
happen now.

No test because it's problematic to test without being root user.