Set correct working directory on startup
When vifm is started it enters into an event loop, as part of which it
changes the working directory to that of the current view's containing
directory. Unfortunately that only happens after an event was received,
i.e., after a key was pressed, for example. That can be a problem if the
working directory is expected to be valid before that (which can be the
case for example inside of a tmux session where a new shell is spawned
in the same directory as the vifm instance).
With this change we add an additional chdir just in front of the event
loop and thereby fix this problem.