Consume input during foreground file operations
Otherwise it remains buffered and is processed after the operation
is over, which might not be the intent of the user.

E.g. try pressing j right after :copy on a large file and then i. It
simply won't work, because at most one character is considered and that
input will actually be processed after the operation is done.

See https://github.com/vifm/vifm/issues/210#issuecomment-469716720