Improve determining widths of strings
Attempt in commit 11dd9071befb227fc63dacd1cbb9cb82b796488d

    Try to improve determining printed widths of chars

    curses outputs some characters via special notation, attempt to
    account for that.

Is not the best one as whether a key is printed as `M-*` or not depends
on encoding used.  So with UTF-8, because of incomplete sequences,
results of `keyname()` might be misleading.

Instead of relying on curses to be correct, translate key sequences
manually using `keyname()` in preview.

This doesn't account for text being printed elsewhere, but this doesn't
produce regressions compared to previous versions either.