Fix buffer over-read in message dialog pane
When processing the last line `ctrl_msg + len + 1U` references memory
past to the original `ctrl_msg`, because it's not clear from `strcspn`
if it found a '\n' (and the string continues) or it reached the end of
the string. In the latter case `skip_char` must not try to skip any
'/'s.