Fix incorrect size for > 4 GiB files on Windows
Got broken in 0.8.1 by commit 0b6574af11d4d1620b6af6f297806c861d1e9db4:

    Display directory size on Windows more correctly

The change wasn't needed for directory sizes.  I guess it was a try of
fix that got committed.

The issue is with overflow in expression `MAXDWORD + 1`, we should
expand the deepest operand to cover the whole outer assignment expression.

Thanks to Allison McNulty (a.k.a. allora).

Fixes #504 on GitHub.