Fix severe performance degradation after ga/gA
Was noticeable after measuring size of large directory trees.  The issue
was caused by comparing timestamp and inode of previously computed value
to modification date and inode of path's parent(!) instead of the path
itself which resulted in recomputation of size.

This was probably exacerbated by taking inode into account in:

    commit 95747a851542a4ed90a1a33524f62ce33b2c7e0d
    Author: xaizek <xaizek@posteo.net>
    Date:   Sun May 24 17:41:08 2020 +0300

        Invalidate cached dir info  on inode change