Don't link with libmagic for AppImage
The library causes trouble because it tries to use pre-compiled magic
data of unsupported version and aborts the application!  Defining
$MAGIC helps, so file of supported format could be distributed in
AppImage, but it might break `file` tool on the host.  An option is
to set $MAGIC temporarily on calling libmagic, but not linking to it
seems simpler.

Thanks to dalvand.

Fixes #749 on GitHub.