Don't reset marking when in dialog mode
Because it shouldn't be dealing with files and dropping marking breaks
operation we're in the middle of.

This fixes #763 on GitHub, but there will be another fix to how macro
expansion is done which will address the issue independently of this
commit, and this change is more of a safety.  A proper solution would
probably be a separation of marking from file view, so it can be passed
where needed rather than preserved across the call chain, which is hard
to do.