Fix %i without " &" breaking some applications
Because it was closing their standard output and error.  There was a
lame implementation and some issues with applications started within
Vifm being closed on Vifm's exit.  With current way of running
background commands, this doesn't seem to happen and that code isn't
needed at all.

Thanks to Daniel J. Perry (a.k.a. BioBox).

Fixes #823 on GitHub.