Made `compare bycontents` skip reading unused data
There is no need to read file's content if there is no other file of the
same size.

Commit message in e738c521604e49dfeb570dc0c043f7be3fc8b51b is wrong:
optimization implemented in this commit was mentioned in #531 on GitHub,
not the one in that one.