Fix clusterfuck of padding issues in ls-like view
This commit just finishes this up.  There were troubles with padding of
columns, with widths of cells, with drawing padding for current cell in
ls-like and miller views...

Part of the reason is that when the code was changed to highlight full
cell, the change didn't update enough code for highlighting only part
of it leaving something very confusing in the code base.  Should be
better now.

Another thing is that fixed number of columns emphasizes determining
sizes of columns much more than when they have varying widths, so code
had to become more precise for this.

Does this mean there are no more bugs with this?  Unlikely.  But at
least the code is great?  I with...  Still had to do special casing for
fixed column count.  Too tired of dealing with this to see it through.