Fix padding overflowing in tls with fixed cols
This is a follow up to 2ef4166bcb0b2008dedce6c8127118b5a3487d0b:

    Fix clusterfuck of padding issues in ls-like view

    [...]

Remove special casing for fixed-column lsview and remove condition on
printing trailing padding cell.  col_width and width_left are not in a
perfect alignment which caused weird behaviour that showed up depending
on the view width.

I think this simplification didn't break anything, but who knows...

Mind that the issues with this are partly caused by my desire to
highlight to the end of the view instead of leaving one empty column of
padding there when fixed column count is 2.