Add separate change log for Lua API
The API is experimental and having a change log is useful to have for
adjusting plugins as the implementation changes.  During development of
the API things should stay fairly stable to not cause breakage too
often, but while stabilizing for v1.0 a bunch of breaking changes can
happen at once.  After that all changes will be documented in the
regular ChangeLog.