Fix effectively infinite loop on waiting for a job
This happened when the job has exited right after printing an error.  In
this situation we're in a nested loop for the purpose of displaying a
message dialog which also calls into bg_check().  Because rip_children()
was called unconditionally and bg_jobs set to NULL during bg_check(), it
did rip the process, but couldn't record it leaving it in a permanent
"running" state with an error thread going crazy because select()
returning immediately but the job was not changing its state.

The issue was there for a long time (since rip_children() was introduced
as a replacement for SIGCHLD), but was hardly visible and it was commit
6e86c8f497f86ca1afaa7ae9c91bf29e2df09d04 which made it very obvious.