Fix vifm.exe using `data/vifmrc` as global config
This is a side-effect of "cleverly" using `data/` as both global config
and application data directory on Windows to simplify handling of global
color schemes.  It's a bad idea because `vifmrc` file in each of those
locations is used differently and results in sample configuration
persisting despite user's attempts to change it in a copy of the sample
vifmrc created in a different location.

Thanks to midrare.

See
https://github.com/vifm/vifm/issues/993#issuecomment-2060237007