Fix event_t pipe filling up and causing a hang
The condition for sending an event was incorrect.  We don't want to do
it when there are any jobs.  Only when there jobs that are handled by
the error thread, otherwise we might end up filling up the pipe while
the thread is looping in import_error_jobs() and doesn't read from it.

An alternative solution would be to mark write end of event pipe as
non-blocking as well, but it seems like a workaround rather than an
actual solution.