Fix processing of unreadable file names and paths
File names that undergo escaping in the UI were processed *after* that
escaping by :rename, :compare, on sorting and in other cases.  This
caused issues as escaped names do not correspond to original files on
the file system.

The cause is commit d4fea5cba81925421f7b2f8e3ee64f99fe594407
("Escape weird characters in file views") which was now done differently
to affect only drawing things in the UI.  As a side-effect of new
implementation, shortened paths in filtered trees or custom views are
now properly escaped as well.

Thanks to jc-SpaceXp.

Fixes #1001 on GitHub.