Refactor handling of macros
It's been over 12 years since refactoring the implementation was
warranted.  This increases number of lines, but reduces amount of
executed code and allows to handle macros better.  Update in the tests
was necessary because now {curr,other}_view are used unconditionally
while previously they were skipped by the check for lack of `%` in the
input string.