Fix loading a global tabs when 'tabscope'="pane"
The code for loading state was originally written for application
startup when global tabs are active by default and only global->pane
scope switch was ever necessary (in cfg/info.c:load_pane()).  With the
introduction of sessions with persistent tabs, pane->global scope switch
also needs to be handled but wasn't until now.

The impact of the bug should be quite low given that users mostly use
one kind of tabs.