Run distcheck in a fresh clone in scripts/deploy
This avoids picking up untracked files in tests/ which is listed in
EXTRA_DIST of Makefile.am as a directory resulting in all of its
contents being included as is.

Makefile.am uses dist-hook to remove some of the files, the top line of
that recipe is still relevant because on distcheck tests/ will contain
build artifacts that shouldn't make it into the archive.