Fix rooted trashes and read-only file-systems
Fix treating a failure to create a trash directory on a read-only
file-system in a Unix environment as a success.

The issue was introduced in commit
dc63ea0c6e26a36431afa971bb9eb624cffc7ec8 which did not differentiate
between two uses of try_create_trash_dir(), that commit was:

    Ignore read-only trash directory error

The issue became apparent on MacOS after commit
4ed5d79ed1a0f3c11fa2f3ad2c1b8b4831f872d1:

     Fix 'trashdir' with "%r" on BSD-like systems

Thanks to feekApp.

Fixes #1079 on GitHub.