Fix unlimited growth of directory histories
Directory history is filled with data read from the state before its
size limit is known.  For this reason it's allowed to grow unlimited and
diverge from the size limit imposed by the 'history' option.  Setting an
option in the configuration to the default value of 15 or not setting it
there at all leaves the directory history size larger than it should be.
Enforce the expected size after processing configuration to avoid the
history growing unbounded.

Fixes the cause of #1075 on GitHub.

Thanks to zoj613.