Get rid of status_t::save_msg_in_list
This doesn't deserve to be part of a global state.

Move implementation of cmd_handler.c:messages_cmd() to ui/statusbar.c to
avoid the need for any global flags or an API that allows to not store a
message in a history (this is the only such case and it can be handled
nicely internally).