Deduplicate associations differently
Instead of dropping programs from the list individually, compare lists
of programs and skip adding duplicates.

In practice duplicated pairs of patterns and commands are rare.  When
they show up, they are likely to be part of an identical list of
commands.  Therefore, we shouldn't be losing anything here.  The
deduplication could even be completely dropped, but probably better to
have it in some form.