Fix ruler size not being updated on menu switches
That is on :cnewer and :colder.

How to reproduce:
 1. :grep if
 2. :find *.cpp
 3. :copen
 4. :colder
 5. G

Update statusbar layout in ui_ruler_set() as well.  Also call
update_window_lazy() instead of wnoutrefresh().

tests/commands/history.c is updated because it ends up calling
ui_ruler_set() which now needs cfg.ruler_format to not be NULL.