Fix buffer cleanup in vifm.vim
Since `:bdelete` doesn't actually remove the buffer, but just makes it
harder to access, it remains recorded as an alternative buffer and
closing the first instance of Vifm ends up processing the same paths
again (because directories are handled on BufEnter event).

Thanks to Ivan Shapovalov (a.k.a., intelfx) and pascualpobil.

Fixes https://github.com/vifm/vifm.vim/issues/95