Workaround sanitizers breaking a Lua test
vifmjob_terminate_running failes to terminate a job.  It kill parent of
the target process, but the process itself keeps running.  On top of
that, while parent ends up being dead no notification about it is
received by Vifm.

This has happened with all three kinds of sanitizers and never happened
without them, so attributing the issue to them.